+++ java/net/meanscript/MInputStream.java
@@ -31,10 +31,10 @@ public int readInt () throws MException
  // bytes:	b[0] b[1] b[2] b[3] b[4] b[5] b[6] b[7]   ...
  // ints:	_________i[0]______|_________i[1]______|_ ...
  int i = 0;
- i |= (readByte() << 24) & 0xff000000;
- i |= (readByte() << 16) & 0x00ff0000;
- i |= (readByte() << 8) & 0x0000ff00;
- i |= (readByte()) & 0x000000ff;
+ i |= (int)((readByte() << 24) & 0xff000000);
+ i |= (int)((readByte() << 16) & 0x00ff0000);
+ i |= (int)((readByte() << 8) & 0x0000ff00);
+ i |= (int)((readByte()) & 0x000000ff);
  return i;
 }
 public void readArray (int [] trg, int numInts) throws MException
